+++ cellxgene_schema_cli/cellxgene_schema/cli.py
@@ -1,6 +1,6 @@
 import click
 import sys
-from cellxgene_schema import validate
+from .validate import validate
 
 
 @click.group(
@@ -33,8 +33,15 @@ def schema_cli():
     default=None,
     type=click.Path(exists=False, dir_okay=False, writable=True),
 )
-def schema_validate(h5ad_file, add_labels_file):
-    if validate.validate(h5ad_file, add_labels_file):
+@click.option(
+    "-v",
+    "--verbose",
+    help="When present will set logging level to debug",
+    is_flag=True
+)
+def schema_validate(h5ad_file, add_labels_file, verbose):
+    is_valid, _ = validate(h5ad_file, add_labels_file, verbose=verbose)
+    if is_valid:
         sys.exit(0)
     else:
         sys.exit(1)