+++ cellxgene_schema_cli/cellxgene_schema/validate.py
@@ -1,5 +1,8 @@
+import logging
 import re
 import sys
+from datetime import datetime
+
 import anndata
 import os
 import pandas as pd
@@ -10,42 +13,14 @@ from numpy import cumprod
 from numpy import isnan
 from numpy import nditer
 from scipy import sparse
-from typing import List, Dict, Union, Optional, Tuple
+from typing import List, Dict, Union, Tuple
 from . import ontology
 from . import schema
 from . import env
 
+logger = logging.getLogger(__name__)
+
 ONTOLOGY_CHECKER = ontology.OntologyChecker()
-
-
-def _curie_remove_suffix(term_id: str, suffix_def: dict) -> Tuple[str, str]:
-    """
-    Remove suffix from a curie term id, if none present return it unmodified
-
-    :param str term_id: the curie term id to validate
-    :param dict{str: list[str], ...} suffix_def: dictionary whose keys are ontology term ids and values
-    are list of allowed suffixes
-
-    :rtype Tuple[str, str]
-    :return the term_id with suffixed stripped, and the suffix
-    """
-
-    id_suffix = ""
-
-    for ontology_name, suffixes in suffix_def.items():
-
-        for suffix in suffixes:
-            suffix = suffix.replace("(", r"\(")
-            suffix = suffix.replace(")", r"\)")
-            search_results = re.search(r"%s$" % suffix, term_id)
-            if search_results:
-                stripped_term_id = re.sub(r"%s$" % suffix, "", term_id)
-                if ONTOLOGY_CHECKER.is_valid_term_id(ontology_name, stripped_term_id):
-                    id_suffix = search_results.group(0)
-
-                    return stripped_term_id, id_suffix
-
-    return term_id, id_suffix
 
 
 class Validator:
@@ -68,6 +43,36 @@ class Validator:
         # Values will be instances of ontology.GeneChecker,
         # keys will be one of ontology.SupportedOrganisms
         self.gene_checkers = dict()
+
+    @staticmethod
+    def _curie_remove_suffix(term_id: str, suffix_def: dict) -> Tuple[str, str]:
+        """
+        Remove suffix from a curie term id, if none present return it unmodified
+
+        :param str term_id: the curie term id to validate
+        :param dict{str: list[str], ...} suffix_def: dictionary whose keys are ontology term ids and values
+        are list of allowed suffixes
+
+        :rtype Tuple[str, str]
+        :return the term_id with suffixed stripped, and the suffix
+        """
+
+        id_suffix = ""
+
+        for ontology_name, suffixes in suffix_def.items():
+
+            for suffix in suffixes:
+                suffix = suffix.replace("(", r"\(")
+                suffix = suffix.replace(")", r"\)")
+                search_results = re.search(r"%s$" % suffix, term_id)
+                if search_results:
+                    stripped_term_id = re.sub(r"%s$" % suffix, "", term_id)
+                    if ONTOLOGY_CHECKER.is_valid_term_id(ontology_name, stripped_term_id):
+                        id_suffix = search_results.group(0)
+
+                        return stripped_term_id, id_suffix
+
+        return term_id, id_suffix
 
     @staticmethod
     def getattr_anndata(adata: anndata.AnnData, attr: str = None):
@@ -102,7 +107,7 @@ class Validator:
             # see https://github.com/theislab/anndata/issues/326#issuecomment-892203924
             self.adata = anndata.read_h5ad(h5ad_path, backed=None)
         except (OSError, TypeError):
-            print(f"Unable to open '{h5ad_path}' with AnnData")
+            logger.info(f"Unable to open '{h5ad_path}' with AnnData")
             sys.exit(1)
 
         self.h5ad_path = h5ad_path
@@ -303,7 +308,7 @@ class Validator:
 
         # Check if there are any allowed suffixes and remove them if needed
         if "suffixes" in curie_constraints:
-            term_id, suffix = _curie_remove_suffix(
+            term_id, suffix = self._curie_remove_suffix(
                 term_id, curie_constraints["suffixes"]
             )
 
@@ -673,7 +678,7 @@ class Validator:
         """
 
         for key, value_def in dict_def["keys"].items():
-
+            logger.debug(f"Validating uns dict for key: {key}")
             if key not in dictionary:
                 if "required" in value_def:
                     self.errors.append(f"'{key}' in '{dict_name}' is not present.")
@@ -729,6 +734,7 @@ class Validator:
 
         # Validate index if needed
         if "index" in self._get_component_def(df_name):
+            logger.debug("Validating index...")
             self._validate_column(
                 pd.Series(df.index),
                 "index",
@@ -738,6 +744,7 @@ class Validator:
 
         # Validate columns
         for column_name in df_definition["columns"].keys():
+            logger.debug(f"Validating column: {column_name}...")
             if column_name not in df.columns:
                 self.errors.append(
                     f"Dataframe '{df_name}' is missing column '{column_name}'."
@@ -933,13 +940,14 @@ class Validator:
                 x = self.adata.X
 
             # Get array without zeros
+            logger.debug("Getting nonzero matrix for type (int) validation...")
+            start = datetime.now()
             non_zeroes_index = x.nonzero()
-
             if max_values_to_check > len(non_zeroes_index[0]):
                 max_values_to_check = len(non_zeroes_index[0])
 
             x_non_zeroes = x[non_zeroes_index[0][:max_values_to_check], non_zeroes_index[1][:max_values_to_check]]
-
+            logger.debug(f"Copy of nonzero matrix created in: {datetime.now() - start}")
             # If all values are zeros then is raw, otherwise if a single value is not an int then return is not raw
             if x_non_zeroes.size < 1:
                 self._raw_layer_exists = True
@@ -1144,11 +1152,12 @@ class Validator:
         self._check_column_availability()
 
         # Checks sparsity
+        logger.debug("Validating sparsity...")
         self._validate_sparsity()
 
         # Checks each component
         for component, component_def in self.schema_def["components"].items():
-
+            logger.debug(f"Validating component: {component}")
             # Skip if component does not exist: only useful for adata.raw.var
             if self.getattr_anndata(self.adata, component) is None:
                 if "required" in component_def:
@@ -1166,10 +1175,10 @@ class Validator:
                 self._validate_embedding_dict()
             else:
                 raise ValueError(f"Unexpected component type '{component['type']}'")
-
         # Checks for raw only if there are no errors, because it depends on the
         # existence of adata.obs["assay_ontology_term_id"] and adata.obs["X_normalization"]
         if not self.errors and "raw" in self.schema_def:
+            logger.debug("Validating raw layer...")
             self._validate_raw()
         else:
             self.warnings.append(
@@ -1188,12 +1197,14 @@ class Validator:
         :return True if successful validation, False otherwise
         :rtype bool
         """
-
+        logger.info("Starting validation...")
         # Re-start errors in case a new h5ad is being validated
         self.errors = []
 
         if h5ad_path:
+            logger.debug("Reading the h5ad file...")
             self._read_h5ad(h5ad_path)
+            logger.debug("Successfully read the h5ad file")
 
         # Fetches schema def from anndata if schema version is not found in AnnData, this fails
         self._set_schema_def()
@@ -1204,12 +1215,14 @@ class Validator:
         # Print warnings if any
         if self.warnings:
             self.warnings = ["WARNING: " + i for i in self.warnings]
-            print(*self.warnings, sep="\n")
+            for w in self.warnings:
+                logger.warning(w)
 
         # Print errors if any
         if self.errors:
             self.errors = ["ERROR: " + i for i in self.errors]
-            print(*self.errors, sep="\n")
+            for e in self.errors:
+                logger.error(e)
             self.is_valid = False
         else:
             self.is_valid = True
@@ -1217,363 +1230,39 @@ class Validator:
         return self.is_valid
 
 
-class AnnDataLabelAppender:
-    """
-    From valid h5ad, handles writing a new h5ad file with ontology/gene labels added
-    to adata.obs and adata.var respectively as indicated in the schema definition
-    """
-
-    def __init__(self, validator: Validator):
-        """
-        From a list of ids and defined constraints, creates a mapping dictionary {id: label, ...}
-
-        :param Validator validator: a Validator object, it's used to get adata and schema defintion for its validation,
-        it's also used to make sure the validation on this adata was successful.
-        """
-
-        if not validator.is_valid:
-            raise ValueError(
-                "AnnData object is not valid or hasn't been run through validation. "
-                "Validate AnnData first before attempting to write labels"
-            )
-
-        if validator.adata.isbacked:
-            self.adata = validator.adata.to_memory()
-        else:
-            self.adata = validator.adata.copy()
-
-        self.validator = validator
-        self.schema_def = validator.schema_def
-        self.errors = []
-        self.was_writing_successful = False
-
-    def _merge_dicts(self, dict1: dict, dict2: dict) -> dict:
-
-        """
-        Recursively merges two dicts, designed to be used to flatten a column definition.
-
-        :params dict dict1: first dict
-        :params dict dict2: second dict
-
-        :rtype dict
-        :return the merged dict
-        """
-
-        merged_dict = dict1.copy()
-
-        for key, value_2 in dict2.items():
-
-            if key == "rule":
-                continue
-
-            if key not in dict1:
-                merged_dict[key] = value_2
-            else:
-                value_1 = dict1[key]
-
-                if not type(value_1) == type(value_2):
-                    raise ValueError("Inconsistent types, impossible to merge")
-
-                if isinstance(value_2, str):
-                    if key == "error_message_suffix":
-                        merged_dict[key] = value_1 + " " + value_2
-                        continue
-                    if not value_2 == value_1:
-                        raise ValueError(
-                            f"Strings types in dependencies cannot be different, {value_1} and {value_2}"
-                        )
-
-                elif isinstance(value_2, list):
-                    merged_dict[key] = list(set(value_1 + value_2))
-                elif isinstance(value_2, dict):
-                    merged_dict[key] = self._merge_dicts(value_1, value_2)
-                else:
-                    raise ValueError(f"merging {type(value_2)} is not implemented")
-
-        return merged_dict
-
-    def _flatten_column_def_with_dependencies(self, column_def: dict) -> dict:
-
-        """
-        Flattens a column definition  that has dependencies, it essentially concatenates all the definitions of the
-        dependencies into the on definition
-
-        :params dict column_def: the column definition that has dependencies in it
-
-        :rtype dict
-        :return the flatten column definition
-        """
-
-        # Do nothing if ther are no dependencies
-        if "dependencies" not in column_def:
-            return column_def
-
-        flatten = column_def.copy()
-        del flatten["dependencies"]
-
-        for dep in column_def["dependencies"]:
-            flatten = self._merge_dicts(flatten, dep)
-
-        return flatten
-
-    def _get_mapping_dict_curie(
-        self, ids: List[str], curie_constraints: dict
-    ) -> Dict[str, str]:
-
-        """
-        From defined constraints it creates a mapping dictionary of ontology IDs and labels.
-
-        :param list[str] ids: Ontology IDs use for mapping
-        :param list[str] curie_constraints: curie constraints e.g.
-        schema_def["components"]["obs"]["columns"]["cell_type_ontology_term_id"]["curie_"]
-
-        :return a mapping dictionary: {id: label, ...}
-        :rtype dict
-        """
-
-        mapping_dict = {}
-        allowed_ontologies = curie_constraints["ontologies"]
-
-        # Remove any suffixes if any
-        # original_ids will have untouched ids which will be used for mapping
-        # id_suffixes will save suffixes if any, these will be used to append to labels
-        # ids will have the ids without suffixes
-        original_ids = ids.copy()
-        id_suffixes = [""] * len(ids)
-
-        if "suffixes" in curie_constraints:
-            for i in range(len(ids)):
-                ids[i], id_suffixes[i] = _curie_remove_suffix(
-                    ids[i], curie_constraints["suffixes"]
-                )
-
-        for original_id, id, id_suffix in zip(original_ids, ids, id_suffixes):
-            # If there are exceptions the label should be the same as the id
-            if "exceptions" in curie_constraints:
-                if original_id in curie_constraints["exceptions"]:
-                    mapping_dict[original_id] = original_id
-                    continue
-
-            for ontology_name in allowed_ontologies:
-                if ontology_name == "NA":
-                    continue
-                if ONTOLOGY_CHECKER.is_valid_term_id(ontology_name, id):
-                    mapping_dict[original_id] = (
-                        ONTOLOGY_CHECKER.get_term_label(ontology_name, id) + id_suffix
-                    )
-
-        # Check that all ids got a mapping. All ids should be found if adata was validated
-        for id in original_ids:
-            if id not in mapping_dict:
-                raise ValueError(f"Add labels error: Unable to get label for '{id}'")
-
-        return mapping_dict
-
-    def _get_mapping_dict_feature_id(self, ids: List[str]) -> Dict[str, str]:
-
-        """
-        Creates a mapping dictionary of gene/feature IDs and labels.
-
-        :param list[str] ids: Gene/feature IDs use for mapping
-
-        :return a mapping dictionary: {id: label, ...}
-        :rtype dict
-        """
-
-        mapping_dict = {}
-
-        for i in ids:
-            organism = ontology.get_organism_from_feature_id(i)
-            mapping_dict[i] = self.validator.gene_checkers[organism].get_symbol(i)
-
-        return mapping_dict
-
-    def _get_mapping_dict_feature_reference(
-        self, ids: List[str]
-    ) -> Dict[str, Optional[ontology.SupportedOrganisms]]:
-
-        """
-        Creates a mapping dictionary of gene/feature IDs and NCBITaxon curies
-
-        :param list[str] ids: Gene/feature IDs use for mapping
-
-        :return a mapping dictionary: {id: label, ...}
-        :rtype dict
-        """
-
-        mapping_dict = {}
-
-        for i in ids:
-            organism = ontology.get_organism_from_feature_id(i)
-            mapping_dict[i] = organism.value
-
-        return mapping_dict
-
-    def _get_labels(
-        self,
-        component: str,
-        column: str,
-        column_definition: dict,
-        label_type: dict,
-    ) -> pd.Categorical:
-
-        """
-        Retrieves a new column (pandas categorical) with labels based on the IDs in 'column' and the logic in the
-        'column_definition'
-
-        :param str component: what dataframe in self.adata to work with
-        :param str column: Column in self.adata with IDs that will be used to retrieve values
-        :param dict column_definition: schema definition of the column
-        e.g. schema_def["obs"]["columns"]["cell_type_ontology_term_id"]
-        :param dict label_type: the type of label
-
-        :rtype pandas.Categorical
-        :return new pandas column with labels corresponding to input column
-        """
-
-        # Set variables for readability
-        current_df = Validator.getattr_anndata(self.adata, component)
-
-        if column == "index":
-            original_column = pd.Series(current_df.index)
-            original_column.index = current_df.index
-        else:
-            original_column = getattr(current_df, column)
-
-        ids = getattr(current_df, column).drop_duplicates().tolist()
-
-        # Flatten column definition (will do so if there are dependencies in the definition
-        column_definition = self._flatten_column_def_with_dependencies(
-            column_definition
-        )
-
-        if label_type == "curie":
-
-            if "curie_constraints" not in column_definition:
-                raise ValueError(
-                    f"Schema definition error: 'add_lables' with type 'curie' was found for '{column}' "
-                    "but no curie constraints were found for the lables"
-                )
-
-            mapping_dict = self._get_mapping_dict_curie(
-                ids, column_definition["curie_constraints"]
-            )
-
-        elif label_type == "feature_id":
-            mapping_dict = self._get_mapping_dict_feature_id(ids=ids)
-
-        elif label_type == "feature_reference":
-            mapping_dict = self._get_mapping_dict_feature_reference(ids=ids)
-
-        else:
-            raise TypeError(
-                f"'{label_type}' is not supported in 'add-labels' functionality"
-            )
-
-        new_column = original_column.copy().replace(mapping_dict).astype("category")
-
-        return new_column
-
-    def _add_column(self, component: str, column: str, column_definition: dict):
-
-        """
-        Adds a new column (pandas categorical) to a component of adata with labels based on the IDs
-        in 'column' and the logic in the 'column_def'
-
-        :param str component: what dataframe in self.adata to work with
-        :param str column: Column in self.adata with IDs that will be used to retrieve values
-        :param dict column_definition: schema definition of the column
-        e.g. schema_def["obs"]["columns"]["cell_type_ontology_term_id"]
-
-        :rtype None
-        """
-
-        for label_def in column_definition["add_labels"]:
-
-            new_column = self._get_labels(
-                component, column, column_definition, label_def["type"]
-            )
-            new_column_name = label_def["to_column"]
-
-            # The sintax below is a programtic way to access obs and var in adata:
-            # adata.__dict__["_obs"] is adata.obs
-            self.adata.__dict__["_" + component][new_column_name] = new_column
-
-    def _add_labels(self):
-
-        """
-        From a valid (per cellxgene's schema) adata, this function adds to self.adata ontology/gene labels
-        to adata.obs and adata.var respectively
-        """
-
-        for component in ["obs", "var"]:
-
-            # Doing it for columns
-            for column, column_def in self.schema_def["components"][component][
-                "columns"
-            ].items():
-
-                if "add_labels" in column_def:
-                    self._add_column(component, column, column_def)
-
-            # Doing it for index
-            index_def = self.schema_def["components"][component]["index"]
-            if "add_labels" in index_def:
-                self._add_column(component, "index", index_def)
-
-    def write_labels(self, add_labels_file: str):
-
-        """
-        From a valid (per cellxgene's schema) h5ad, this function writes a new h5ad file with ontology/gene labels added
-        to adata.obs  and adata.var respectively
-
-        :param str add_labels_file: Path to new h5ad file with ontology/gene labels added
-
-        :rtype None
-        """
-
-        # Add labels in obs
-        self._add_labels()
-
-        # Write file
-        try:
-            self.adata.write_h5ad(add_labels_file, compression="gzip")
-        except Exception as e:
-            self.errors.append(f"Writing h5ad was unsuccessful, got exception '{e}'.")
-
-        # Print errors if any
-        if self.errors:
-            self.errors = ["ERROR: " + i for i in self.errors]
-            print(*self.errors, sep="\n")
-            self.was_writing_successful = False
-        else:
-            self.was_writing_successful = True
-
-
-def validate(h5ad_path: Union[str, bytes, os.PathLike], add_labels_file: str = None):
-
+def validate(h5ad_path: Union[str, bytes, os.PathLike], add_labels_file: str = None, verbose: bool = False) -> (
+        bool, list):
+    from .write_labels import AnnDataLabelAppender
     """
     Entry point for validation.
 
     :param Union[str, bytes, os.PathLike] h5ad_path: Path to h5ad file to validate
     :param str add_labels_file: Path to new h5ad file with ontology/gene labels added
 
-    :return True if successful validation, False otherwise
-    :rtype bool
+    :return (True, []) if successful validation, (False, [list_of_errors] otherwise
+    :rtype tuple
     """
 
     # Perform validation
+    start = datetime.now()
+    if verbose:
+        logging.basicConfig(level=logging.DEBUG)
+    else:
+        logging.basicConfig(level=logging.INFO, format="%(message)s")
     validator = Validator()
     validator.validate_adata(h5ad_path)
+    logger.info(f"Validation complete in {datetime.now() - start} with status is_valid={validator.is_valid}")
 
     # Stop if validation was unsuccessful
     if not validator.is_valid:
-        return False
+        return False, validator.errors
 
     if add_labels_file:
+        label_start = datetime.now()
         writer = AnnDataLabelAppender(validator)
         writer.write_labels(add_labels_file)
+        logger.info(f"H5AD label writing complete in {datetime.now() - label_start}, was_writing_successful: {writer.was_writing_successful}") # noqa E501
 
-        return validator.is_valid & writer.was_writing_successful
+        return validator.is_valid and writer.was_writing_successful, validator.errors + writer.errors
 
-    return True
+    return True, validator.errors